refactor(notes): extract use_auto_title hook from NoteDetail
